registry/pyproject.toml
[project]
name = "hikyaku-registry"
version = "0.1.0"
description = "A2A-native message broker and agent registry server"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "a2a-sdk",
    "sse-starlette",
    "redis[hiredis]",
    "pydantic",
    "pydantic-settings",
    "PyJWT[crypto]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/hikyaku_registry"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
